Добавил методы для работы с сайтом

+++ src/main/kotlin/icfpc/y2023/db/repository/ProblemRepository.kt
@@ -3,4 +3,4 @@ package icfpc.y2023.db.repository
 import icfpc.y2023.db.model.Problem
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface ProblemRepository : JpaRepository<Problem, Long>
+interface ProblemRepository : JpaRepository<Problem, Int>
